+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Edit Inventory Instances: Display a Save & keep editing button. Refs UIIN-2457.
 * Edit Inventory Holdings: Display a Save & keep editing button. Refs UIIN-2404.
 * "holdings-storage" API version upgrade. Refs UIIN-2926.
+* Edit Inventory Items: Display a Save & keep editing button. Refs UIIN-2456.
 * Import facets and the function for building a search query from `stripes-inventory-components`.
 
 ## [11.0.4](https://github.com/folio-org/ui-inventory/tree/v11.0.4) (2024-04-30)